making icc happy 2

+++ tests/test-compressQ.C
@@ -75,7 +75,7 @@ int main(int argc, char** argv)
 	A[16+16*N] = 1;
 	deg[0]  = 4; deg[1] = 4; deg[2] = 4;deg[3] = 2; deg[4] = 1; deg[5] =2;
 	for (size_t i=0; i<size_t(N); ++i)
-		A[11+i*N] = A[7+i*N] = A[3+i*N] = i % 10;
+		A[11+i*N] = A[7+i*N] = A[3+i*N] = double(i % 10);
 
 	double * B = new double[N*N] ;
 	FFLAS::fcopy(F,N*N,B,1,A,1);
